+++ src/groth16/prover.rs
@@ -469,91 +469,3 @@ where
 
     Ok(proofs)
 }
-
-pub fn create_proof<E, C, P: ParameterSource<E>>(
-    circuit: C,
-    params: P,
-    r: E::Fr,
-    s: E::Fr,
-) -> Result<Proof<E>, SynthesisError>
-where
-    E: Engine,
-    C: Circuit<E> + Send,
-{
-    let proofs =
-        create_proof_batch_priority::<E, C, P>(vec![circuit], params, vec![r], vec![s], false)?;
-    Ok(proofs.into_iter().next().unwrap())
-}
-
-pub fn create_random_proof<E, C, R, P: ParameterSource<E>>(
-    circuit: C,
-    params: P,
-    rng: &mut R,
-) -> Result<Proof<E>, SynthesisError>
-where
-    E: Engine,
-    C: Circuit<E> + Send,
-    R: RngCore,
-{
-    let proofs =
-        create_random_proof_batch_priority::<E, C, R, P>(vec![circuit], params, rng, false)?;
-    Ok(proofs.into_iter().next().unwrap())
-}
-
-pub fn create_proof_batch<E, C, P: ParameterSource<E>>(
-    circuits: Vec<C>,
-    params: P,
-    r: Vec<E::Fr>,
-    s: Vec<E::Fr>,
-) -> Result<Vec<Proof<E>>, SynthesisError>
-where
-    E: Engine,
-    C: Circuit<E> + Send,
-{
-    create_proof_batch_priority::<E, C, P>(circuits, params, r, s, false)
-}
-
-pub fn create_random_proof_batch<E, C, R, P: ParameterSource<E>>(
-    circuits: Vec<C>,
-    params: P,
-    rng: &mut R,
-) -> Result<Vec<Proof<E>>, SynthesisError>
-where
-    E: Engine,
-    C: Circuit<E> + Send,
-    R: RngCore,
-{
-    create_random_proof_batch_priority::<E, C, R, P>(circuits, params, rng, false)
-}
-
-pub fn create_proof_priority<E, C, P: ParameterSource<E>>(
-    circuit: C,
-    params: P,
-    r: E::Fr,
-    s: E::Fr,
-    priority: bool,
-) -> Result<Proof<E>, SynthesisError>
-where
-    E: Engine,
-    C: Circuit<E> + Send,
-{
-    let proofs =
-        create_proof_batch_priority::<E, C, P>(vec![circuit], params, vec![r], vec![s], priority)?;
-    Ok(proofs.into_iter().next().unwrap())
-}
-
-pub fn create_random_proof_priority<E, C, R, P: ParameterSource<E>>(
-    circuit: C,
-    params: P,
-    rng: &mut R,
-    priority: bool,
-) -> Result<Proof<E>, SynthesisError>
-where
-    E: Engine,
-    C: Circuit<E> + Send,
-    R: RngCore,
-{
-    let proofs =
-        create_random_proof_batch_priority::<E, C, R, P>(vec![circuit], params, rng, priority)?;
-    Ok(proofs.into_iter().next().unwrap())
-}
